shopping cart and checkout - base

+++ HalceraAPI.Model/Requests/Price/CreatePriceRequest.cs
@@ -9,13 +9,13 @@ namespace HalceraAPI.Models.Requests.Price
     public class CreatePriceRequest
     {
         [Required]
-        public double? Amount { get; set; }
+        public decimal? Amount { get; set; }
         [Required]
         public Currency? Currency { get; set; }
 
         /// <summary>
         /// indicate if price is discounted
         /// </summary>
-        public double? DiscountAmount { get; set; }
+        public decimal? DiscountAmount { get; set; }
     }
 }
